[Refactor] LiveData 테스트코드 확장함수 따로 분리
